testing_suite/requirements.txt
# Requirements for the AudioMuse-AI Testing & Comparison Suite
requests>=2.28.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
pytest>=7.0.0
pytest-json-report>=1.5.0
pytest-timeout>=2.1.0
